requirements.txt
yfinance
google-api-python-client
youtube-transcript-api
python-dotenv
requests
transformers
torch
sentencepiece
fastapi
uvicorn